[SL-1] Added initial game menu with React
