+++ index.html
@@ -1,13 +1,15 @@
 <!DOCTYPE html>
 <html lang="en">
-  <head>
-    <meta charset="UTF-8">
-    <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <link rel="stylesheet" href="src/style.css">
-  </head>
-  <body>
+<head>
+  <meta charset="UTF-8">
+  <meta name="viewport" content="width=device-width, initial-scale=1.0">
+  <link rel="stylesheet" href="src/style.css">
+</head>
+<body>
+<div id="react-root"></div>
 <div>siema</div>
 <div class="container" id="container"></div>
-    <script src="src/script.js" type="module"></script>
-  </body>
+<script src="src/index.jsx" type="module"></script>
+<script src="src/Game.jsx" type="module"></script>
+</body>
 </html>
