+++ package.json
@@ -1,5 +1,8 @@
 {
   "dependencies": {
+    "@vitejs/plugin-react": "^4.2.1",
+    "react": "^18.3.1",
+    "react-dom": "^18.3.1",
     "three": "0.162.0",
     "vite": "^5.2.7",
     "vite-plugin-top-level-await": "^1.4.1"
